app/src/main/java/com/example/rrulepractice/utils/QuoteManager.kt
package com.example.rrulepractice.utils

import android.content.Context
import com.google.gson.Gson
import java.nio.charset.Charset

class QuoteManager {

    var quoteList = emptyArray<Quote>()
    private var currentQuoteIndex = 0


    fun populateQuoteFromAssets(context: Context, fileName: String) {
        val inputStream = context.assets.open(fileName)
        val size: Int = inputStream.available()
        val buffer = ByteArray(size)
        inputStream.read(buffer)
        inputStream.close()
        val json = String(buffer, Charset.defaultCharset())
        val gson = Gson()
        quoteList = gson.fromJson(json, Array<Quote>::class.java)
    }

    fun populateQuotes(quote: Array<Quote>) {
        quoteList = quote
    }

    fun getCurrentQuote(): Quote {
        return quoteList[currentQuoteIndex]
    }

    fun getNextQuote(): Quote {
        if(currentQuoteIndex == quoteList.size - 1) return quoteList[currentQuoteIndex]
        return quoteList[++currentQuoteIndex]
    }

    fun getPreviousQuote(): Quote {
        if(currentQuoteIndex == 0) return quoteList[currentQuoteIndex]
        return quoteList[--currentQuoteIndex]
    }

}

data class Quote(val quote: String, val author: String)